created basic layout of main.py
Setup of the main.py file. The file imports pygame and calls the init
method of pygame.

Then the window is setup with size, title and the clock object is
instantiated.

Then the main loop is created and there it first gets the delta time,
then calls the input, update, and draw methods.