made the player able to collide with the walls
the loop first takes the neccessary values from the current wall. Then
it checks if the wall is a left wall or right wall. After that the code
(depending on the direction) checks if the player hits the wall, if yes
call the movable_horizontal method with the direction that the player
cant move to. then break the loop