add terrain_type as parameter
A new parameter was added to the throw method. This parameter takes a
string and then the program checks if it has the value of either
"roof/ground" or "wall". This is done to enable different behaivor
depending on where the exiting portal is placed.

If the portal is a roof portal the y speed wants to be kept

If the portal is a wall portal the y speed should be turned into x speed